+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 0.13.1
 
+- fixed adding providers/agencies modal for subscriptions
 - fixed start page typo
 
 0.13